WIP
Fixed some AX.25 seq number state managements

Start on NET/ROM networking

+++ src/main/java/net/tarpn/packet/impl/ax25/IFrame.java
@@ -26,8 +26,8 @@ public class IFrame extends BaseAX25Packet implements AX25Packet.InformationFram
       AX25Call destCall,
       AX25Call sourceCall,
       Command command,
-      int sendSeqNumber,
-      int recvSeqNumber,
+      byte sendSeqNumber,
+      byte recvSeqNumber,
       boolean pollFinalSet,
       Protocol protocol,
       byte[] info) {
